fix(ma): harden SI mapping — security, performance, data integrity
- Remove hardcoded JWT_SECRET and DATABASE_URL defaults (config.py)
- Add prod/staging runtime validation for missing secrets
- Replace client JWT forwarding with service-to-service token (SSRF fix)
- Apply RBAC via require_role() on all SI mapping endpoints
- Optimize N+1 flush loop → single batch flush in bulk_add_to_buyers
- Convert float → Decimal for revenue/transaction_value/coefficient models
- Add LIKE escape for special chars in KSIC search
- Add max_length=100 to BulkAddBuyersRequest.si_company_ids
- Rename _get_jwt_secret → get_jwt_secret (public API convention)
- Improve migration 035 downgrade with batch operations
- Add CSV header validation and error path tests

+++ deal-mgmt/app/models/io_transaction.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from decimal import Decimal
+
 from sqlalchemy import ForeignKey, Integer, Numeric, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -29,4 +31,4 @@ class IOTransaction(Base):
         index=True,
     )
     target_io_name: Mapped[str | None] = mapped_column(String(200), nullable=True)
-    transaction_value: Mapped[float] = mapped_column(Numeric(20, 2), nullable=False)
+    transaction_value: Mapped[Decimal] = mapped_column(Numeric(20, 2), nullable=False)
